Add files via upload
In this project, I worked with data from the entertainment industry. I studied a dataset with records on movies and shows. The research will focus on the “Golden Age” of television, which began in 1999 with the release of The Sopranos and is still ongoing.

The aim of this project is to investigate how the number of votes a title receives from IMDb users impacts its ratings. The assumption is that highly-rated shows (focused on TV shows, ignoring movies) released during the “Golden Age” of television also have the most votes.